test: move dial test to testing package
