--- internal/jujuclient/package_test.go
@@ -1,13 +1,0 @@
-// Copyright 2024 Canonical.
-
-package jujuclient_test
-
-import (
-	"testing"
-
-	gc "gopkg.in/check.v1"
-)
-
-func TestPackage(t *testing.T) {
-	gc.TestingT(t)
-}
