+++ testing/dial_test.go
@@ -1,6 +1,6 @@
 // Copyright 2025 Canonical.
 
-package jujuclient_test
+package testing
 
 import (
 	"context"
